Extracted CurrentBlock from ProgressBar
